Final fixes

+++ client/src/pages/dashboard.js
@@ -4,14 +4,14 @@ import { React, useState, useEffect, useRef, } from 'react';
 import Table from 'react-bootstrap/Table';
 import { useNavigate } from 'react-router-dom';
 import Footer from "../components/footer";
-import { Link, Paper, TableBody, TableCell, TableContainer, TableHead, TableRow } from "@mui/material";
+import { Paper, TableBody, TableCell, TableContainer, TableHead, TableRow } from "@mui/material";
 import Department from "../components/department";
-// import dotenv from "dotenv"
 
 function Dashboard() {
     // const { auth }= useSelector(state=>state.auth);
-  const modalRefDepartment = useRef();
-  const navigate = useNavigate();
+    console.log("in dashboard");
+    const modalRefDepartment = useRef();
+    const navigate = useNavigate();
     const [result, setResults] = useState([]);
     useEffect(() => {
         axios.get("http://localhost:5000/api/v1/department/show")
@@ -24,33 +24,42 @@ function Dashboard() {
                 setResults(data.data.results);
             });
     }, []); // Empty dependency array means it will only run once on component mount
-   const handleDeleteDepartment = async (id) => {
-    axios.delete(`department?_id=${id}`)
-    .then((res) => {
-        if (res.status === 200) {
-            console.log("sucess in dep", res);
-            // setResults(res);
-            alert("action done sucessfully");
-            navigate("/dashboard");
-        } else {
-            // setError((res.response).toString());
-            console.log("in else case error");
-        }
-    })
-    .catch(err => {
-        console.log("err", err.response.data.message);
-        // setError((err.response.data.message).toString());
-        // setOpen(true)
+    // if i ass result variable in [] then its render automatically continuosly 
+    const handleDeleteDepartment = async (id) => {
+        const confirmed = window.confirm('Are you sure you want to delete this department?');
+        if (confirmed) {
 
-    })
-    console.log("dep delete successfull", id);
-   }
-   const handleUpdateDepartment = async (id) => {
-    console.log("dep delete successfull", id);
-    localStorage.setItem("depId", id);
-    // modalRefDepartment.current.id= id;
-    modalRefDepartment.current.departmentOpen();
-   }
+            axios.delete(`department?_id=${id}`)
+                .then((res) => {
+                    if (res.status === 200) {
+                        console.log("sucess in dep", res);
+                        // setResults(res);
+                        alert("action done sucessfully");
+                        navigate("/dashboard");
+                    } else {
+                        // setError((res.response).toString());
+                        console.log("in else case error");
+                    }
+                })
+                .catch(err => {
+                    console.log("err", err.response.data.message);
+                    // setError((err.response.data.message).toString());
+                    // setOpen(true)
+
+                })
+            console.log("dep delete successfull", id);
+        } else {
+            console.log("delete cancel successfull", id);
+            navigate("/dashboard")
+        }
+    }
+    const handleUpdateDepartment = async (row) => {
+
+        console.log("dep update successfull", row);
+        localStorage.setItem("rowData", JSON.stringify(row));
+        // modalRefDepartment.current.id= id;
+        modalRefDepartment.current.departmentOpen();
+    }
     return (
         <>
             <div style={{ height: "750px" }}>
@@ -62,7 +71,7 @@ function Dashboard() {
                 <p>Futhermore please signup or signin to continue</p>
             </div>} */}
                 <h1 style={{ color: "green" }}>Well Come to your Dashboard page</h1>
-                <p style={{ color: "greenyellow", margin:"5px" }}><strong>help the students to chose your university for themby adding more details about your university
+                <p style={{ color: "greenyellow", margin: "5px" }}><strong>help the students to chose your university for themby adding more details about your university
                 </strong>
                 </p>
                 <div style={{ background: "white", padding: "10px 20px" }}>
@@ -73,15 +82,15 @@ function Dashboard() {
                             <Table className={"table"} aria-label="simple table" size={"small"} stickyHeader={true} >
                                 <TableHead>
                                     <TableRow>
-                                        <TableCell align="right">department</TableCell>
-                                        <TableCell align="right">semester</TableCell>
-                                        <TableCell align="right">fee</TableCell>
-                                        <TableCell align="right">applyMerit</TableCell>
+                                        <TableCell align="right">Department</TableCell>
+                                        <TableCell align="right">Semester</TableCell>
+                                        <TableCell align="right">Fee</TableCell>
+                                        <TableCell align="right">ApplyMerit</TableCell>
                                         <TableCell align="right">Level</TableCell>
-                                        <TableCell align="right">openingDate</TableCell>
-                                        <TableCell align="right">closingDate</TableCell>
+                                        <TableCell align="right">Opening Date</TableCell>
+                                        <TableCell align="right">Closing Date</TableCell>
                                         <TableCell align="right">Delete</TableCell>
-                                        <TableCell align="right">update</TableCell>
+                                        <TableCell align="right">Update</TableCell>
                                     </TableRow>
                                 </TableHead>
                                 <TableBody>
@@ -95,10 +104,10 @@ function Dashboard() {
                                             <TableCell align="right">{row.openingDate}</TableCell>
                                             <TableCell align="right">{row.closingDate}</TableCell>
                                             <TableCell align="right">
-                                            <button style={{ backgroundColor: "#d11a2a", color: "white" }} onClick={() => handleDeleteDepartment(row._id)}>Delete</button>
+                                                <button style={{ backgroundColor: "#d11a2a", color: "white" }} onClick={() => handleDeleteDepartment(row._id)}>Delete</button>
                                             </TableCell>
                                             <TableCell align="right">
-                                            <button style={{ backgroundColor: "#FFE569", color: "black" }} onClick={() => handleUpdateDepartment(row._id)}>Update</button>
+                                                <button style={{ backgroundColor: "#FFE569", color: "black" }} onClick={() => handleUpdateDepartment(row)}>Update</button>
                                             </TableCell>
                                         </TableRow>
                                     ))}
